+++ classes/class-lsx-customizer-woocommerce.php
@@ -23,7 +23,10 @@ if ( ! class_exists( 'LSX_Customizer_WooCommerce' ) ) {
 			add_filter( 'body_class', array( $this, 'body_class' ), 2999 );
 
 			add_action( 'template_redirect', array( $this, 'thankyou_page' ), 2999 );
-			add_action( 'lsx_entry_inside_top', array( $this, 'checkout_steps' ), 15 );
+
+			if ( empty( get_theme_mod( 'lsx_two_step_checkout', false ) ) ) {
+				add_action( 'lsx_entry_inside_top', array( $this, 'checkout_steps' ), 15 );
+			}
 
 			add_action( 'wp', array( $this, 'cart_extra_html' ), 2999 );
 			add_action( 'wp', array( $this, 'checkout_extra_html' ), 2999 );
@@ -71,6 +74,9 @@ if ( ! class_exists( 'LSX_Customizer_WooCommerce' ) ) {
 				'priority'    => 1,
 			) ) );
 
+			/**
+			 * Checkout Layout
+			 */
 			$wp_customize->add_setting( 'lsx_wc_checkout_layout', array(
 				'default' => 'default',
 				'sanitize_callback' => array( $this, 'sanitize_select' ),
@@ -99,6 +105,49 @@ if ( ! class_exists( 'LSX_Customizer_WooCommerce' ) ) {
 				'0' => esc_html__( 'Default', 'lsx-customizer' ),
 			);
 
+			/**
+			 * Distraction Free Checkout
+			 */
+			$wp_customize->add_setting( 'lsx_distraction_free_checkout', array(
+				'sanitize_callback' => array( $this, 'sanitize_checkbox' ),
+			) );
+
+			$wp_customize->add_control( new WP_Customize_Control(
+				$wp_customize,
+				'lsx_distraction_free_checkout',
+				array(
+					'label'       => esc_html__( 'Distraction Free Checkout', 'lsx-customizer' ),
+					'description' => esc_html__( 'Removes all clutter from the checkout, allowing the customer to focus entirely on that procedure. Removes the stepped cart and checkout.', 'lsx-customizer' ),
+					'section'     => 'lsx-wc-checkout',
+					'settings'    => 'lsx_distraction_free_checkout',
+					'type'        => 'checkbox',
+					'priority'    => 3,
+				)
+			) );
+
+			/**
+			 * Two Step Checkout
+			 */
+			$wp_customize->add_setting( 'lsx_two_step_checkout', array(
+				'sanitize_callback' => array( $this, 'sanitize_checkbox' ),
+			) );
+
+			$wp_customize->add_control( new WP_Customize_Control(
+				$wp_customize,
+				'lsx_two_step_checkout',
+				array(
+					'label'       => esc_html__( 'Two Step Checkout', 'lsx-customizer' ),
+					'description' => esc_html__( 'Separates the customer details collection form, and the order summary / payment details form in to two separate pages. Removes the stepped cart and checkout.', 'lsx-customizer' ),
+					'section'     => 'lsx-wc-checkout',
+					'settings'    => 'lsx_two_step_checkout',
+					'type'        => 'checkbox',
+					'priority'    => 4,
+				)
+			) );
+
+			/**
+			 * Thank you page options
+			 */
 			$pages = get_pages();
 
 			foreach ( $pages as $key => $page ) {
@@ -111,7 +160,7 @@ if ( ! class_exists( 'LSX_Customizer_WooCommerce' ) ) {
 				'section'     => 'lsx-wc-checkout',
 				'settings'    => 'lsx_wc_checkout_thankyou_page',
 				'type'        => 'select',
-				'priority'    => 3,
+				'priority'    => 5,
 				'choices'     => $choices,
 			) ) );
 
@@ -125,7 +174,7 @@ if ( ! class_exists( 'LSX_Customizer_WooCommerce' ) ) {
 				'description' => esc_html__( 'Extra HTML to display at checkout page (bottom/right).', 'lsx-customizer' ),
 				'section'     => 'lsx-wc-checkout',
 				'settings'    => 'lsx_wc_checkout_extra_html',
-				'priority'    => 4,
+				'priority'    => 6,
 				'type'        => 'wysiwyg',
 			) ) );
 
@@ -255,13 +304,23 @@ if ( ! class_exists( 'LSX_Customizer_WooCommerce' ) ) {
 		 * @since 1.1.1
 		 */
 		public function body_class( $classes ) {
+			$distraction_free = get_theme_mod( 'lsx_distraction_free_checkout', false );
+			$two_step_checkout = get_theme_mod( 'lsx_two_step_checkout', false );
 			if ( is_checkout() ) {
 				$layout = get_theme_mod( 'lsx_wc_checkout_layout', 'default' );
 
-				if ( 'stacked' === $layout ) {
+				if ( 'default' === $layout ) {
+					$classes[] = 'lsx-wc-checkout-layout-default';
+				} elseif ( 'stacked' === $layout ) {
 					$classes[] = 'lsx-wc-checkout-layout-stacked';
 				} elseif ( 'columns' === $layout ) {
 					$classes[] = 'lsx-wc-checkout-layout-two-column-addreses';
+				}
+				if ( ! empty( $distraction_free ) ) {
+					$classes[] = 'lsx-wc-checkout-distraction-free';
+				}
+				if ( ! empty( $two_step_checkout ) ) {
+					$classes[] = 'lsx-wc-checkout-two-steps';
 				}
 			}