+++ changelog.md
@@ -4,8 +4,16 @@
 
 ### Added
 
+- Adding color palette options for the customizer.
+- Added Distraction free checkout option.
+- Added 2 steps checkout option.
+
+### Updated
+
+- Updated Checkout and Cart styles.
 - Added customizer colour options for the Block editor palette
 - Added integration for the [WooCommerce Grid / List toggle](https://wordpress.org/plugins/woocommerce-grid-list-toggle/) plugin.
+
 
 ### Changed